Move graph state up to MainContainer
