+++ src/Flowchart.tsx
@@ -1,14 +1,10 @@
 import { useState, useRef, useCallback, useMemo } from 'react';
 import ReactFlow, {
   ReactFlowProvider,
-  addEdge,
-  useNodesState,
-  useEdgesState,
   Controls,
   ReactFlowInstance,
-  Connection,
   Node,
-  Edge,
+  Edge
 } from 'reactflow';
 import 'reactflow/dist/style.css';
 
@@ -22,57 +18,11 @@ import CacheNode from './nodeTypes/CacheNode.js';
 import ClientNode from './nodeTypes/ClientNode.js';
 import LoadBalancerNode from './nodeTypes/LoadBalancerNode.js';
 
-const initialNodes: Node[] = [
-  {
-    id: 'dndnode_1',
-    type: 'client',
-    data: { label: 'Client' },
-    position: {
-      x: 250,
-      y: 5,
-    },
-  },
-  {
-    id: 'dndnode_2',
-    type: 'server',
-    data: { label: 'Server' },
-    position: {
-      x: 273,
-      y: 92.25,
-    },
-  },
-  {
-    id: 'dndnode_3',
-    type: 'database',
-    data: { label: 'Database' },
-    position: {
-      x: 280.5,
-      y: 196.25,
-    },
-  },
-];
-
-const initialEdges: Edge[] = [
-  {
-    source: 'dndnode_1',
-    sourceHandle: null,
-    target: 'dndnode_2',
-    targetHandle: null,
-    id: 'client-to-server',
-  },
-  {
-    source: 'dndnode_2',
-    sourceHandle: null,
-    target: 'dndnode_3',
-    targetHandle: null,
-    id: 'server-to-db',
-  },
-];
-
 let id = 4;
 const getId = () => `dndnode_${id++}`;
 
-const Flowchart = () => {
+type FlowchartProps  = {nodes: Node[], edges: Edge[], onNodesChange: any, onEdgesChange: any, onConnect: any, setNodes: any};
+const Flowchart = ({nodes, edges, setNodes, onNodesChange, onEdgesChange, onConnect}: FlowchartProps) => {
   const nodeTypes = useMemo(() => ({
     custom: CustomNode,
     server: ServerNode,
@@ -81,13 +31,8 @@ const Flowchart = () => {
     client: ClientNode,
     loadbalancer: LoadBalancerNode}), []);
   const reactFlowWrapper = useRef(null);
-  const [nodes, setNodes, onNodesChange] = useNodesState(initialNodes);
-  const [edges, setEdges, onEdgesChange] = useEdgesState(initialEdges);
   const [reactFlowInstance, setReactFlowInstance] = useState<ReactFlowInstance | null>(null);
 
-  const onConnect = useCallback((params: Connection) => {
-    setEdges((eds) => addEdge(params, eds));
-  }, []);
 
   const onDragOver = useCallback((event: React.DragEvent) => {
     event.preventDefault();
@@ -96,7 +41,7 @@ const Flowchart = () => {
 
   const setCustomData = (id: string, newCustomData: string) => {
     // Traverse nodes, find the right ID, then update its data, then setNodes to the new Nodes object
-    setNodes(prevNodes => {
+    setNodes((prevNodes: Node[]) => {
       return prevNodes.map( node => {
         if (node.id === id) {
           const data = {...node.data, customData: newCustomData};
@@ -134,7 +79,7 @@ const Flowchart = () => {
         data: { label: `${type}`, setCustomData },
       };
 
-      setNodes((nds) => nds.concat(newNode));
+      setNodes((nds: Node[]) => nds.concat(newNode));
     },
     [reactFlowInstance]
   );
